64-bit work in progress
git-svn-id: https://svn.code.sf.net/p/osirix/code/osirix@1000 4e9e6e3d-4551-47eb-9dbe-a6127f9400a1

+++ SeriesView.h
@@ -43,8 +43,8 @@
 
 - (id)initWithFrame:(NSRect)frame seriesRows:(int)rows  seriesColumns:(int)columns;
 
-- (int)tag;
-- (void)setTag:(int)theTag;
+- (long)tag;
+- (void)setTag:(long)theTag;
 - (NSMutableArray *)imageViews;
 - (void)setImageViewMatrixForRows:(int)rows  columns:(int)columns;
 - (void)updateImageTiling:(NSNotification *)note;
